Model and Route setup
Here I created the models for Trips, Days, and Events. I also created and tested the post, get , and delete routes for trips. I have not worked with updating trips yet. Not sure if we need Day and Event controllers yet, so that is not set up. I also created the file needed to establish the pool connection in order to make requests.